% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFeatures-family.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{annotateFeatures}
\usage{
annotateFeatures(XCMSnExp, ChemFormulaParam, filter = TRUE)
}
\arguments{
\item{XCMSnExp}{XCMS object that contains a list of features (ie. after
running \code{groupChromPeaks()}).}

\item{ChemFormulaParam}{A ChemFormulaParam object that contains the lists of
molecular formulas and adducts to annotate the features.}

\item{filter}{Logical. Whether to remove those features that can't be matched
to a SOI.}
}
\description{
Annotate features using a list of molecular formulas and adducts.
}
