% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemFormulaParam-class.R
\docType{class}
\name{ChemFormulaParam-class}
\alias{ChemFormulaParam-class}
\alias{ChemFormulaParam}
\title{ChemFormulaParam}
\description{
Main ionic formula structure
}
\section{Slots}{

\describe{
\item{\code{DB}}{Character, path to the molecular formula database}

\item{\code{adlist}}{Character, list of adducts that will be used to generate ionic
formulas}

\item{\code{ppm}}{Numeric, instrumental mass error (in parts per million, ppm).}

\item{\code{ionFormulas}}{Dataframe. NOT to be manually provided by the user. It is
automatically filled when the object is created, using the provided DB and
adlist.}
}}

