% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSOI.R
\name{consensusSOI}
\alias{consensusSOI}
\title{consensusSOI}
\usage{
consensusSOI(RHermesExp, SOIids, minSOI = NULL, rtwin = 5)
}
\arguments{
\item{RHermesExp}{An RHermesExp object containing the SOI lists}

\item{SOIids}{Which SOI lists are to be grouped.}

\item{minSOI}{Minimum number of SOI lists where a particular SOI has to be
present in order to be preserved in the final list.}

\item{rtwin}{Retention time tolerance to match the SOIs from the different
lists.}
}
\description{
Group a set of SOI lists into a consensus list that contains all
SOIs common in a subset of them. The concept is similar to the grouping step
from peaks to features present in other tools.
}
\details{
consensusSOI is commonly used when multiple quality control samples
are available in order to reduce the number of SOIs and ensure only
consistent signals are included into the inclusion list.
}
\examples{
if(FALSE){
 #Find SOIs common in at least two of the three selected SOI lists
 consensusSOI(RHermesExp, 1:3, 2) 
}
}
