% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSOI-family.R
\name{fastSOI}
\alias{fastSOI}
\title{fastSOI}
\usage{
fastSOI(MSnExp, ChemFormulaParam, minint = 1000, peakwidth = c(5, 30))
}
\arguments{
\item{MSnExp}{An MSnExp object containing one or multiple MS1 files}

\item{ChemFormulaParam}{A ChemFormulaParam object, containing all ionic
formulas to be annotated in the data.}

\item{minint}{Minimum intensity of the scans used in the calculated XICs.
Defaults to 1000.}
}
\description{
A new implementation of SOI detection algorithm that avoids the
datapoint abstraction.
}
\details{
The algorithm is faster than the regular SOI detection used in
RHermes, but on the other hand loses the computational commodity of having
all experimental datapoints annotated.
}
