% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSOI-family.R
\name{fastSOIfromList}
\alias{fastSOIfromList}
\title{fastSOIfromList}
\usage{
fastSOIfromList(MSnExp, RHermesExp, SOI_id = 1, RTtol = 3, tol = 3, thr = 1000)
}
\arguments{
\item{MSnExp}{An MSnExp object containing one or multiple MS1 files}

\item{RHermesExp}{RHermes object that contains a SOI (Scans of Interest)
list.}

\item{SOI_id}{Index of the SOI that will be used as reference for the matching.}

\item{RTtol}{Numeric. Retention time shift tolerance in seconds.}

\item{tol}{Numeric. Gap (datapoints with <thr intensity) tolerance.}

\item{thr}{Numeric. Intensity threshold for signal detection.}
}
\description{
Quantify SOIs in multiple files using an RHermes SOI list as a
template.
}
\details{
The function uses the fastSOI approach to quickly detect SOIs in
multiple files using a target list of all SOIs in a given RHermesExp
SOIList.
}
