% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSOI-family.R
\name{filterSOIByIL}
\alias{filterSOIByIL}
\title{filterSOIByIL}
\usage{
filterSOIByIL(RHermesExp, SOI_id, IL_id)
}
\arguments{
\item{RHermesExp}{RHermes object that contains a SOI (Scans of Interest)
list.}

\item{SOI_id}{Index of the SOI that will be used as reference for the matching.}

\item{IL_id}{Numeric. Index of the inclusion list that will be used to filter the SOIs.}
}
\description{
Filter a SOI list and keep only those SOIs that were included in
the inclusion list.
}
\details{
The rationale of this function is that one usually wants to quantify
only those signals for which he has acquired MS2 data.
}
