% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFeatures}
\alias{plotFeatures}
\title{plotFeatures}
\usage{
plotFeatures(XCMSnExp, id, log = TRUE)
}
\arguments{
\item{XCMSnExp}{An XCMSnExp object with processed features (ie. after running
\code{groupChromPeaks})}

\item{id}{\code{numeric}: The number (row index) of the feature to plot.}
}
\description{
\code{plotFeatures} creates a plot of the XIC profiles of the peaks
grouped in a single feature.
}
\seealso{
plotSummarizedExp
}
