% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSOI.R
\name{fastSOIfromList}
\alias{fastSOIfromList}
\title{fastSOIfromList}
\usage{
fastSOIfromList(MSnExp, struct, SOI_id = 1, rtwin = 3, tol = 3, thr = 1000)
}
\description{
Quantify SOIs in multiple files using an RHermes SOI list as a
template.
}
\details{
The function uses the fastSOI approach to quickly detect SOIs in
multiple files using a target list of all SOIs in a given RHermesExp
SOIList.
}
